+++ pkg/app/app.go
@@ -21,6 +21,7 @@ type Config struct {
 	ConfigVersion       string
 	DisableForceUpgrade bool
 	Name                string
+	PauseReconciliation bool
 	UserConfigMapName   string
 	UserSecretName      string
 }
@@ -36,6 +37,9 @@ func NewCR(c Config) *applicationv1alpha1.App {
 		}
 		if !c.DisableForceUpgrade {
 			annotations["chart-operator.giantswarm.io/force-helm-upgrade"] = "true"
+		}
+		if c.PauseReconciliation {
+			annotations[annotation.AppOperatorPaused] = "true"
 		}
 	}